requirements.txt
word2number
sentence-transformers
python-dotenv
google-generativeai
transformers